+++ packages/design-system/src/components/bottom-modal/handlers/label-handler.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from '../../../providers/theme-provider';
+import { useTheme } from '../../../providers/ThemeProvider';
 import { AppTheme } from '../../../hooks/useAppThemeSetup';
 import {
   BottomSheetHandle,
