+++ packages/design-system/src/components/colors/color-picker/color-picker.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo } from 'react';
 import { Pressable, StyleSheet, View } from 'react-native';
 import { Text } from 'react-native-paper';
 import { AppTheme } from '../../../hooks/useAppThemeSetup';
-import { useTheme } from '../../../providers/theme-provider';
+import { useTheme } from '../../../providers/ThemeProvider';
 import { SelectOption } from '../../select-buttons/select-buttons';
 import { useBottomModal } from '../../../hooks/useBottomModal';
