+++ packages/design-system/src/components/label-switch/label-switch.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 import { Pressable, StyleProp, TextStyle, ViewStyle } from 'react-native';
 import { Switch, Text } from 'react-native-paper';
 import { AppTheme } from '../../hooks/useAppThemeSetup';
-import { useTheme } from '../../providers/theme-provider';
+import { useTheme } from '../../providers/ThemeProvider';
 
 const getStyle = (theme: AppTheme) => {
   return {
