+++ packages/design-system/src/components/screen-wrapper/screen-wrapper.tsx
@@ -8,7 +8,7 @@ import {
 } from 'react-native';
 import { EdgeInsets, useSafeAreaInsets } from 'react-native-safe-area-context';
 import { AppTheme } from '../../hooks/useAppThemeSetup';
-import { useTheme } from '../../providers/theme-provider';
+import { useTheme } from '../../providers/ThemeProvider';
 
 const getStyles = ({
   theme,
