+++ packages/design-system/src/components/select-items/select-items.tsx
@@ -11,7 +11,7 @@ import {
 } from 'react-native';
 import { Button, HelperText, MD3Theme, Searchbar } from 'react-native-paper';
 import { useScreenWidth } from '../../hooks/useScreenWidth';
-import { useTheme } from '../../providers/theme-provider';
+import { useTheme } from '../../providers/ThemeProvider';
 import { BREAKPOINTS } from '../select-buttons/select-buttons';
 
 const logger = getLogger('SelectItems');
