+++ packages/design-system/src/hooks/useConfirm.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 import {
   ConfirmContext,
   ConfirmContextType,
-} from '../providers/confirm-provider';
+} from '../providers/ConfirmProvider';
 
 export const useConfirm = (): ConfirmContextType => {
   const context = useContext(ConfirmContext);
